fatops: Only write out remaining data if it's there
fat_file_close incorrectly assumed that there is always data to be written
in the buffer of a file which has been opened for writing, resulting in the
duplication of the last 256 bytes of any file whose length is an exact
multiple of 256.

Fixed by actually using the dirty flag of the buffer that was meant for
this case.