fastloader: Split enum definition and storage
fastloader.h defines an enum and its storage variable at the same time
which would result in a compiler error in fastloader.c because the enum
is redefined there. The error was hidden by another problem in fastloader-ll.h
which caused fastloader.h to not define anything if fastloader-ll.h was
included before.