iec: Add another experimental delay
sd2iec was missing close commands (0xe0-0xef) on the bus when a program
tried to read the first two bytes of every file shown in the directory.

Adding a 60us delay fixed the problem, but the exact time and the
correctness of this fix has not been investigated yet.