Track number of write buffers
The code did not track the number of buffers open for writing at all,
so after reading the error channel or closing one write buffer the
dirty LED was always turned off. This patch tracks the number of write
buffers in the upper four bits of active_buffers and changes the LED
state in free_buffer automagically.