Transmit actual file length if known
This patch changes the value in the low byte of the next line link
pointer of the directory listing to (filesize MOD 254)+2 if the
exact file size can be easily determined (i.e. only on FAT for now).
This enables a directory-parsing program to calculate the exact file
size and also to easily distinguish if the value was known or not.