doscmd: Fully emulate B-R/B-W
Almost every documentation mentions that B-R and B-W should not
be used because they are somehow defective - so they were just
mapped to U1/U2 before.

The game Pirates! actually relies on the original behaviour of B-R
which sets the length of the buffer to the value of the first byte of
the sector.

B-W is also patched to set the first byte to the position of the
buffer read/write pointer, but has not been tested yet.