Use external tool for the bootloader signature
Remove the bootloader signature structure in main.c and change
the build process to use an external tool for adding the signature
to the binary. This way the added space is filled with 0xff instead
of 0x00 and the linker won't complain about overlapping sections
with the mega32 builds anymore.

crcgen-new.c can be found at http://snowcat.de/sd2iec/crcgen-new.c