ff: Invalidate the buffer in check_fs
check_fs reads sector 0 directly because move_fs_window ignores any
attempts to read it. It needs to set the current sector of the buffer
it uses to 0, otherwise move_fp_window will assume that its previous
contents are still valid and skip reading the requested sector if it
is the same as the previous one before the call to check_fs.

This problem can appear when mounting a partitioned drive twice in
a row, the first mount will read the boot sector into the buffer
and the second mount will see the contents of the MBR instead of the
contents of the boot sector, causing the mount operation to fail.