Replace all char variables by uint8_t
This patch changes all char variables in the code to uint8_t.
As sd2iec was always compiled with -funsigned-char this does not
change the function of the code, although there is one place where
it was possible to remove a local variable.

The string.h functions are wrapped within ustring.h as they are
defined using char * and require type casts to work with uint8_t.