sdcard: Improve sector read performance
This patch reduces the time to read a single sector from ~8400us
(data CRCs enabled) or ~6700us (data CRCs disabled) to ~5100us
(CRC) or ~5050us (no CRC) with my main test card. The improvement
is achieved by interleaving the SPI access in the main data transfer
loop.

Previously it would send a dummy byte, wait for the completed
transfer, read the received data and then calculate the CRC and store
the byte. With this patch the order of those operations within the loop
changes to wait/read/write, thus using the data transmission delay to
calculate the CRC and store the byte in the buffer.