FatFs: Volume Labels are not LFNs
A part of trace_path errorneously considered any entry with the
AM_VOL bit set to be part of a long file name. Because of that
deleting (and probably also renaming) a file whose LFN directory
entries are preceded by a volume label would not work. Instead
of deleting the LFN and the 8.3 name the code deleted the
volume label and all LFN entries, but left the 8.3 name in the
directory.