sdcard: Always try SD_SEND_OP_COND
Move the block that sends SD_SEND_OP_COND to its own function and
always call it. Previously the initialisation of Ver. 1.x SD cards
would not reach this block, thus violating the specification.

It is still not 100% correct as the HCS bit in ACMD41 should be
set to 0 for those cards, but I'll wait and see how it works out in
the real world.