sdcard: Send GO_IDLE to both cards simultaneously
Send the GO_IDLE_STATE to both cards simultaneously if CONFIG_TWINSD
is enabled because after power-on the cards are in MMC/SD mode instead
of SPI and card 1 can interfere with the initialisation of card 0
because card 1 wasn't in SPI mode yet.

It is unclear why there were no conflicts before the pin assignments
were changed.