Don't free system buffers unless required
This patch adds the function free_all_user_buffers and changes most
users of free_all_buffers to the new function. Previously free_all_buffers
would've freed the system buffers holding the BAM of a mounted D64 image
resulting in a crash the next time the BAM of a still-mounted image is
accessed.