Use dummy callbacks in allocated buffers
Set the callback functions of just-allocation buffers to a dummy
function that just returns success so the callers don't need to
check if a callback has been set.

Saves two bytes of flash compared to the previous version, but
that is missing at least one callback-check (resulting in a crash
when Turbodisk is used to load a file from a D64 image).