Don't write headers for known image file types
Avoid writing headers and don't mangle the file name if
a file is written that is potentially mountable by sd2iec.

Without this patch a disk->D64 tool would create a file on the
card that isn't mountable by sd2iec when the extension mode
specifies x00 headers or file type extensions.