Free buffers on error in d64ops, m2iops
Free the current buffer if an error occurs in d64ops and m2iops
(fatops already did it). This avoids some rare cases of buffer
leakage and fixes the problem that filling up the free space in a
D64 image extends the image file to about 1.1MB because the refill
callback was still called and tried to write the data to sector 0/0.