iec.c: Do not clear the error channel if secondary 15 is closed
Do not force an OK message into the error channel when secondary #15
is closed. The assumption recorded in commit 8a0e9350 that a 1541
resets its error channel when secondary 15 is closed is incorrect
and was caused by a flawed test program that used INPUT#1,A$ to
partially read the error channel which actually reads all of it
but returns only the error number.