iec: Check for ATN during initialisation
Check the state of the ATN line right after initialising the ports
and set DATA to low if ATN is active. This will cause a
C64 that tries to access the serial bus to hang until sd2iec
is fully initialized and able to answer the request.

The port initialisation has been reordered to reduce the time
between port initialisation and the ATN check to a minimum,
reducing the race condition window if the boot loader has
already set DATA low.