FatFs: Fix random corruption on FAT32 partitions
When a new directory entry is created, f_open calls sync(fs) to
force writing of the changed directory sector and later reads
FSBUF.sect to remember the directory sector in the FIL structure.

However, when FSINFO sector support is enabled, sync() will change
FSBUF.sect to zero if it needs to rewrite the FSINFO sector.
This zero ends up as the directory sector in the FIL structure,
so whenever the directory sector needs to be updated, move_window
will only flush the current buffer contents to disk but will
not change the current buffer sector. Because of this the updated
directory information will end up in a random sector, usually
a sector in the FAT area in tests.

Fixed by moving the sync(fs) call to the common read/write code
after the assignment of fp->dir_sect. Although the sync call
is not required when a file is opened only for reading it won't hurt
because the current buffer contents are not marked dirty at that point.