Fix read-after-EOI behaviour
For most file types reading from a channel that has sent a byte with
EOI the Commodore drives will not return any data at all, resulting
in an ST value of 66 (EOI plus read-timeout). sd2iec would return
more data bytes in this situation, fixed by resetting the buf->read
status to 0 which will result in the same behaviour as original drives.

Since this fix must not be applied for direct channels ('#'), use
largebuffer_refill for all # channels and rename it to
directbuffer_refill to improve code clarity.