sdcard: Use proper timeouts instead of counting attempted transfers
Use proper 500ms timeouts in two places instead of counting
the number of bytes read from the card that didn't have the
expected value. As the SD card specification requires that every
command completes in 250ms or less the chosen timeout is safe and
will work with any SPI/CPU frequency.