Move fatcluster/realname to a pvt union in cbmdirent_t
Move the filename and realname fields to a union of ops-specific,
private information in the cbmdirent_t structure. A new field
opstype is used to remember which part of the union is valid, used
whenever the code previously checked for "fatcluster != 0" to
see if the cbmdirent_t referred to a file on FAT.

Moving the name field to the front of the structure saved about
100 bytes of flash in addition to the removal of some useless
checks and smaller code for "opstype == OPSTYPE_FAT" (1 byte
comparison) instead of "opstype != 0" (4 byte comparison).