d64ops: Reduce implicit parameter passing
Add a dxx branch in the private struct of cbmdirent_t that
holds the track/sector/entry number of the directory entry of
the file. It's used in d64ops.c the remove the previously implicit
parameters in a few global variables that were set as a side
effect of the usual order of calls in other pieces of code. The
new version can now just read the entry based on the data
in cbmdirent_t, side effects of previous calls are not
relied upon anymore.