Change chdir semantics
Previously chdir would directly manipulate the current directory
in partitions[...].current_dir. This patch removes this behaviour
and lets chdir only update the directory in the path_t object it's
called with. This allows parse_path to call chdir instead of
manipulating the FAT cluster in its local path_t itself and
forces every caller of chdir() that really wants to change
the current directory to update partitions[...].current_dir.

This change will simplify the implementation of disk image
formats that offer subdirectories because the path_t object
can now (mostly) be handled as an opaque object outside of
the ops implementation.