Check that the target dir is empty in rmdir
Add a check in rmdir to see if there are any entries in the directory
that is to be deleted. Although the same check already happens in
FatFs it can still trigger here if disk image formats with their own
subdirectories are implemented.

To simplify these checks without loss of generality the . and ..
entries of FAT subdirectories are now completely hidden from the
user.