Enable remote display for uIEC, larsp and sw2 and RTC for sw2
Turn on remote display support on the uIEC, larsp and sw2 configurations.
This required defining I2C/IntRQ pins on uIEC (PD0/PD1/PD2) and
an IntRQ pin for larsp (PC7, already reserved for this purpose on FE3).

I2C-RTC support is enabled on sw2 too, it was already on for larsp.