d64ops: Fix raw directory for D64/D71/D81
The current directory entries in path_t point to the first sector
with actual entries for D64/D71/D81, so reading the directory in
raw mode must start in sector 0 for those image formats instead.

It's not possible to use the BAM sector for those formats as
current directory - although it would remove many DNP special cases,
the original drives completely ignore the link pointer in sector 0
of the directory track unless the raw directory is read.