Ensure that SRQ is always driven even on devices that don't connect it
Modify the IEC output functions to enable a pullup on the SRQ line on
devices that use the same pins for IEC input and output to avoid
spurious signals on that pin when burst mode support is added later.

This patch also moves the "generic" IEC output functions into the AVR
section because the change is not generic and lpc17xx has always used
its own output functions.