LPC17xx: Use timer capture interrupts instead of GPIO for IEC
Use the capture interrupt of the two IEC timers on LPC17xx instead of
the GPIO interrupts for ATN and CLOCK to allow for more flexibility
in pin selection because some of the capture pins of the LPC cannot
generate GPIO interrupts.

Because the timers of the LPC will not generate capture interrupts
when stopped timeout functionality was moved to a third timer for now,
although an implementation based on matching instead of stopping may
be possible.