i2c_lpc17xx: Rewrite to allow arbitrary chained block transfers
Rewrite the I2C code for the LPC17xx with chained block transfers as
the basic operation and reimplement the other operations based on
that. The old functions were incapable of accessing chips that require
more than a single byte for register selections, the new ones can
transmit an arbitrary number of bytes before switching to receive
mode.

Implementation of the same chained transfers for AVR is possible,
but was left out for now.