fatops: Check for error in fat_file_close when flushing
Check if buf->refill() has signalled an error when flushing any
remaining data in fat_file close and stop processing if it has. If
buf->refill() signals an error it has already called f_close on the
file, calling it again in fat_file_close overwrites the actual error
message with "DRIVE NOT READY" because the first f_close made sure
that the FIL object cannot be reused.