fastloader.c: Clear rxflag later in dolphin_putc
Move the parallel_clear_rxflag call in dolphin_putc behind the
parallel_write call to ignore glitches on the handshake line.

In a setup with relatively high-capacity parallel cables short
(40-80ns) glitches were observed on the handshake line when the data
lines were switched from 0xff to a value with many 0 bits, causing a
skipped byte because sd2iec saw that glitch as an incoming handshake.