Add an "RTC multiplexer" for RTC module selection at runtime
Add a small "multiplexer" similiar to the diskmux that is auto-enabled
when more than one RTC module is selected to choose one RTC at
run-time. The selection is based on the hardcoded call order in rtc.c
to use the best RTC that is found.

DS1307 and DS3231 are still mutually exclusive because they share the
same source code file with minor compile-time switches for each of the
two chips.