Prepare lpc176x version for bootloader support
Prepare the lpc176x version for bootloader support by moving the
start of the main program to 0x4000 and prepending a small trampoline
code that changes the interrupt vector base and jumpts to the
main code. A bootloader info struct is embedded twice in the binary,
once at 0x100 to allow quick file checks and once at 0x4100 for
checking the version of the in-flash copy.