Kill HW_NAME define
Remove the HW_NAME define that was used to set distinct version
message prefixes for uIEC and everything else. People seem to test
only for one of the prefixes in their software, leaving the other half
of the user base out in the cold.

The new, common prefix is "SD2IEC" as it is the name of the software
project.